% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{carousel}
\alias{carousel}
\title{Carrousel in r}
\usage{
carousel(..., number_div = 3, width = "100\%", height = "100\%")
}
\arguments{
\item{...}{div_carousel functions}

\item{number_div}{number of future div_carousel}

\item{width}{width}

\item{height}{height}
}
\description{
Carrousel in r
}
