% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{connect_db}
\alias{connect_db}
\title{Connect to the good database}
\usage{
connect_db(dev = golem::app_dev())
}
\arguments{
\item{dev}{TRUE or FALSE}
}
\value{
connect
}
\description{
Connect to the good database
}
