% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{div_carousel}
\alias{div_carousel}
\title{div_carousel}
\usage{
div_carousel(..., active = FALSE)
}
\arguments{
\item{...}{tagList}

\item{active}{if TRUE, show up in the carousel}
}
\description{
div_carousel
}
