% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{refresh}
\alias{refresh}
\title{refresh table}
\usage{
refresh(id, id_refresh_msg, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{id input}

\item{id_refresh_msg}{id refrech message}

\item{session}{session}
}
\description{
refresh table
}
