% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app_css_bs4dash}
\alias{run_app_css_bs4dash}
\title{Run CSS bs4dash Shiny Application}
\usage{
run_app_css_bs4dash(...)
}
\arguments{
\item{...}{A series of options to be used inside the app.}
}
\description{
Run CSS bs4dash Shiny Application
}
