% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app_database}
\alias{run_app_database}
\title{Run database Shiny Application}
\usage{
run_app_database(...)
}
\arguments{
\item{...}{A series of options to be used inside the app.}
}
\description{
Run database Shiny Application
}
