% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app_shinipsum}
\alias{run_app_shinipsum}
\title{Run shinispum Shiny Application}
\usage{
run_app_shinipsum(...)
}
\arguments{
\item{...}{A series of options to be used inside the app.}
}
\description{
Run shinispum Shiny Application
}
