% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app_with_p_f}
\alias{run_app_with_p_f}
\title{Run non-blocking Shiny Application}
\usage{
run_app_with_p_f(...)
}
\arguments{
\item{...}{A series of options to be used inside the app.}
}
\description{
Run non-blocking Shiny Application
}
